Revert "move some tests to contrib ops + cosmetic changes"
This reverts commit 4cda9297e257a6f6b902724e8113bf5d5a62df29.
